fix: 수익률 계산 버그 수정 (MVP 완성)

+++ src/main/java/lotto/domain/Wallet.java
@@ -51,7 +51,7 @@ public class Wallet {
         long totalPrice = ranks.stream()
             .mapToLong(RankPrice::getPrice)
             .sum();
-        return totalPrice / originalMoney;
+        return (double)totalPrice / originalMoney;
     }
 
     public int getRankCount(int rank) {
